moved authorize middleware for tables to folder
