+++ lib/middleware/authorize/authorizeProjects.js
@@ -1,4 +1,4 @@
-const Project = require('../models/Project');
+const Project = require('../../models/Project');
 
 module.exports = async (req, res, next) => {
   try {
@@ -8,6 +8,7 @@ module.exports = async (req, res, next) => {
     next();
   } catch (error) {
     error.message = 'You are not authorized to modify this project';
+    error.status = 403;
     next(error);
   }
 };
