+++ lib/middleware/authorize/authorizeUsers.js
@@ -1,4 +1,4 @@
-import User from '../models/User';
+const User = require('../../models/User');
 
 module.exports = async (req, res, next) => {
   try {
@@ -7,7 +7,8 @@ module.exports = async (req, res, next) => {
     if (userId !== user.userId) throw new Error();
     next();
   } catch (error) {
-    error.message = 'You are not authorized to delete this user';
+    error.message = 'You are not authorized to modify this user';
+    error.status = 403;
     next(error);
   }
 };
